Refractored QR-Scanner und Player Element mit Auslagerung in verschiedene Datein für bessere readability und maintainability.
Buttons auf mobile größer gemacht und Zeit Leiste entfernt (hat nicht so viel Sinn ergeben)
Modi funktionieren soweit.

Der Logger wird effizenter verwendet um error der verschiedenen Bereiche des Players zu loggen.
