+++ web-app/src/components/pages/PlayPage/Authorised/Player/controller/types.ts
@@ -1,0 +1,18 @@
+export type PlayerElementControllerOptions = {
+    currentTrackId: string | null;
+    onError?: (message: string) => void;
+    startAtSeconds?: number;
+    stopAtSeconds?: number | null;
+    startAtMiddle?: boolean;
+    startAtRandom?: boolean;
+    minDistanceFromEndSeconds?: number;
+    maxPlayDurationSeconds?: number | null;
+};
+
+export type PlayerElementControllerState = {
+    isPlaying: boolean;
+    loading: boolean;
+    autoPausedLocked: boolean;
+    progressPct: number;
+    togglePlay: () => Promise<void>;
+};
