+++ web-app/src/components/pages/PlayPage/Authorised/QRScanner/index.ts
@@ -1,0 +1,2 @@
+export { QrScannerView as QrScanner } from "./components/QrScannerView";
+export { default } from "./components/QRScannerModal";
