+++ web-app/src/components/pages/PlayPage/Authorised/QRScanner/styles/QRScannerElement.module.css
@@ -1,0 +1,25 @@
+.scanner {
+    position: absolute;
+    inset: 0;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    overflow: hidden;
+}
+
+.reader {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+}
+
+.reader :global(video),
+.reader :global(canvas) {
+    width: 100%;
+    height: 100%;
+}
